requirements.txt
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
scipy==1.7.3
streamlit==1.22.0
Pillow==2.2.1